requirements.txt
python-binance
python-dotenv
python-telegram-bot==13.15
requests
backports.zoneinfo; python_version < "3.9"
numpy
ta-lib-binary
pyinstaller
